[INTERNAL] Revert "sap.ui.fl Remove '.Component' suffix logic"
This reverts commit 73895741a95b554459f78a5442905dea752591d6.

Reason for revert: Aligning build with backend capabilities and will be reintroduced later.

CR-Id: 002075125800001080392023
BCP: n/a

+++ src/sap.ui.fl/src/sap/ui/fl/write/_internal/SaveAs.js
@@ -10,6 +10,7 @@ sap.ui.define([
 	"sap/ui/fl/Utils",
 	"sap/base/Log",
 	"sap/ui/fl/Layer",
+	"sap/base/util/includes",
 	"sap/base/util/merge",
 	"sap/base/util/restricted/_omit",
 	"sap/ui/fl/registry/Settings"
@@ -21,6 +22,7 @@ sap.ui.define([
 	Utils,
 	Log,
 	Layer,
+	includes,
 	merge,
 	_omit,
 	Settings
@@ -133,31 +135,60 @@ sap.ui.define([
 	}
 
 	function _getDirtyDescrChanges(vSelector) {
-		var oFlexControllerPersistence = ChangesController.getFlexControllerInstance(vSelector)._oChangePersistence;
-		if (!oFlexControllerPersistence) {
-			return [];
+		var oDescriptorFlexControllerPersistence = ChangesController.getDescriptorFlexControllerInstance(vSelector)._oChangePersistence;
+		if (oDescriptorFlexControllerPersistence) {
+			var aDescrChanges = oDescriptorFlexControllerPersistence.getDirtyChanges();
+			aDescrChanges = aDescrChanges.slice();
+			return aDescrChanges;
 		}
-		return oFlexControllerPersistence.getDirtyChanges().filter(function(oChange) {
-			return DescriptorChangeTypes.getChangeTypes().includes(oChange.getChangeType());
-		});
+		return [];
 	}
 
-	function _getDirtyChanges(vSelector) {
+	function _getDirtyUIChanges(vSelector) {
 		var oFlexControllerPersistence = ChangesController.getFlexControllerInstance(vSelector)._oChangePersistence;
-		if (!oFlexControllerPersistence) {
-			return [];
+		if (oFlexControllerPersistence) {
+			var aUIChanges = oFlexControllerPersistence.getDirtyChanges();
+			aUIChanges = aUIChanges.slice();
+			return aUIChanges;
 		}
-		var aChanges = oFlexControllerPersistence.getDirtyChanges();
-		aChanges = aChanges.slice();
-		return aChanges;
+		return [];
+	}
+
+	function _arePersistenciesTheSame(vSelector) {
+		var oFlexControllerPersistence = ChangesController.getFlexControllerInstance(vSelector)._oChangePersistence;
+		var oDescriptorFlexControllerPersistence = ChangesController.getDescriptorFlexControllerInstance(vSelector)._oChangePersistence;
+		// If the base application is already an app variant, the references and both persistences are same
+		return oFlexControllerPersistence === oDescriptorFlexControllerPersistence;
+	}
+
+	function _reactOnChangesBasedOnPersistences(vSelector, bArePersistencesEqual, oAppVariant) {
+		var aDescrChanges = [];
+		if (bArePersistencesEqual) {
+			_getDirtyDescrChanges(vSelector).forEach(function(oChange) {
+				// UI and Descriptor changes need to be separated here so as to perform different operations on changes
+				if (includes(DescriptorChangeTypes.getChangeTypes(), oChange.getChangeType())) {
+					aDescrChanges.push(oChange);
+				} else {
+					_moveChangesToNewFlexReference(oChange, oAppVariant);
+				}
+			});
+		} else {
+			_getDirtyUIChanges(vSelector).forEach(function(oChange) {
+				_moveChangesToNewFlexReference(oChange, oAppVariant);
+			});
+
+			aDescrChanges = _getDirtyDescrChanges(vSelector);
+		}
+
+		return aDescrChanges;
 	}
 
 	function _deleteDescrChangesFromPersistence(vSelector) {
 		// In case of app variant, both persistences hold descriptor changes and have to be removed from one of the persistences
 		_getDirtyDescrChanges(vSelector).forEach(function(oChange) {
-			if (DescriptorChangeTypes.getChangeTypes().includes(oChange.getChangeType())) {
+			if (includes(DescriptorChangeTypes.getChangeTypes(), oChange.getChangeType())) {
 				// In case there are UI changes, they will be sent to the backend in the last resolved promise and will be removed from the persistence
-				ChangesController.getFlexControllerInstance(vSelector)._oChangePersistence.deleteChange(oChange);
+				ChangesController.getDescriptorFlexControllerInstance(vSelector)._oChangePersistence.deleteChange(oChange);
 			}
 		});
 	}
@@ -182,6 +213,7 @@ sap.ui.define([
 		saveAs: function(mPropertyBag) {
 			var oAppVariantClosure;
 			var oAppVariantResultClosure;
+			var bArePersistencesEqual = false;
 
 			return AppVariantFactory.prepareCreate(mPropertyBag)
 				.then(function(oAppVariant) {
@@ -192,15 +224,8 @@ sap.ui.define([
 					return _setTransportAndPackageInfoForAppVariant(oAppVariantClosure, oTransportInfo);
 				})
 				.then(function() {
-					var aDescrChanges = [];
-					_getDirtyChanges(mPropertyBag.selector).forEach(function(oChange) {
-						// UI and Descriptor changes need to be separated here so as to perform different operations on changes
-						if (DescriptorChangeTypes.getChangeTypes().includes(oChange.getChangeType())) {
-							aDescrChanges.push(oChange);
-						} else {
-							_moveChangesToNewFlexReference(oChange, oAppVariantClosure);
-						}
-					});
+					bArePersistencesEqual = _arePersistenciesTheSame(mPropertyBag.selector);
+					var aDescrChanges = _reactOnChangesBasedOnPersistences(mPropertyBag.selector, bArePersistencesEqual, oAppVariantClosure);
 					return _getInlineChangesFromDescrChanges(aDescrChanges);
 				})
 				.then(function(aAllInlineChanges) {
@@ -216,15 +241,22 @@ sap.ui.define([
 				})
 				.then(function(oResult) {
 					oAppVariantResultClosure = merge({}, oResult);
-					_deleteDescrChangesFromPersistence(mPropertyBag.selector);
+
+					if (bArePersistencesEqual) {
+						_deleteDescrChangesFromPersistence(mPropertyBag.selector);
+					}
 
 					var oFlexController = ChangesController.getFlexControllerInstance(mPropertyBag.selector);
 
-					var aUIChanges = _getDirtyChanges(mPropertyBag.selector); //after removing descr changes, all remaining dirty changes are UI changes
+					var aUIChanges = _getDirtyUIChanges(mPropertyBag.selector);
 					if (aUIChanges.length) {
 						// Save the dirty UI changes to backend => firing PersistenceWriteApi.save
 						return oFlexController.saveAll(ChangesController.getAppComponentForSelector(mPropertyBag.selector), true)
 							.catch(function(oError) {
+								if (bArePersistencesEqual) {
+									_deleteDescrChangesFromPersistence(mPropertyBag.selector);
+								}
+
 								// Delete the inconsistent app variant if the UI changes failed to save
 								return this.deleteAppVariant({
 									id: mPropertyBag.id
@@ -239,6 +271,10 @@ sap.ui.define([
 					return Promise.resolve();
 				}.bind(this))
 				.then(function() {
+					//Reference Application Usecase: Since the UI changes have been successfully saved, the descriptor inline changes will now be removed from persistence
+					if (!bArePersistencesEqual) {
+						_deleteDescrChangesFromPersistence(mPropertyBag.selector);
+					}
 					return oAppVariantResultClosure;
 				})
 				.catch(function(oError) {
@@ -279,7 +315,7 @@ sap.ui.define([
 				.then(function() {
 					var aDescrChanges = [];
 					_getDirtyDescrChanges(mPropertyBag.selector).forEach(function(oChange) {
-						if (DescriptorChangeTypes.getChangeTypes().includes(oChange.getChangeType())) {
+						if (includes(DescriptorChangeTypes.getChangeTypes(), oChange.getChangeType())) {
 							aDescrChanges.push(oChange);
 						}
 					});
